Add handler program timeout and fix header parsing
Fix header parsing -- was working on buffer, but saving dup, which was the whole request.

Call alarm(3) before execlp(3) when starting the handler program

Define _XOPEN_SOURCE 600, so that it builds with clean C99

Fix sample echo_handler, headers are now $REQHEADERS

Update man page to mention the 3 configuration defines and what they do.

Bump to 1.2